"#51": Refactor and cleanup new components
