+++ compiler/src/main/scala/de/zalando/apifirst/generators/AstScalaPlayEnricher.scala
@@ -6,8 +6,6 @@ import de.zalando.apifirst.ScalaName._
 import de.zalando.apifirst.generators.DenotationNames._
 import de.zalando.apifirst.naming.Reference
 
-import scala.annotation.tailrec
-
 /**
   * @author  slasch 
   * @since   21.12.2015.
@@ -57,7 +55,7 @@ trait Transformation[T] extends EnrichmentStep[T] {
   * Enriches AST with information related to Types
   */
 class ScalaPlayTypeEnricher(val app: StrictModel) extends Transformation[Type] with
-  /*TypeValidatorsStep with*/ DataGeneratorsStep with
+  DataGeneratorsStep with
   AliasesStep with TraitsStep with ClassesStep with CommonDataStep {
 
   override val data = app.typeDefs.toSeq
@@ -78,7 +76,7 @@ class ScalaPlayCallEnricher(val app: StrictModel) extends Transformation[ApiCall
   * Enriches AST with information related to Parameters
   */
 class ScalaPlayParameterEnricher(val app: StrictModel) extends Transformation[Parameter]
-  /*with ParameterAliasesStep */with ParametersValidatorsStep with CommonParamDataStep {
+  with ParametersValidatorsStep with CommonParamDataStep {
 
   override def data = app.params.toSeq.map { case (r, p) => r.name -> p }
 
@@ -92,196 +90,6 @@ trait EnrichmentStep[InputType] {
 
   def steps: Seq[SingleStep] = Seq.empty
 
-}
-
-trait CommonParamDataStep extends EnrichmentStep[Parameter] with CommonData {
-
-  override def steps = types +: super.steps
-
-  /**
-    * Puts type information into the denotation table
-    * @return
-    */
-  protected def types: SingleStep = parameter => table => singleParameter(parameter)
-
-  private def singleParameter(paramPair: (Reference, Parameter)) = {
-    val (ref, param) = paramPair
-    val t = param.typeName
-    val name = typeName(t, ref)
-    Map(COMMON -> Map(TYPE_NAME -> name, FIELDS -> fieldsForType(t), MEMBER_NAME -> memberName(t, ref)))
-  }
-
-}
-trait CommonCallDataStep extends EnrichmentStep[ApiCall] with CommonData {
-
-  override def steps = types +: super.steps
-
-  /**
-    * Puts type information into the denotation table
-    * @return
-    */
-  protected def types: SingleStep = parameter => table => singleParameter(parameter)
-
-  private def singleParameter(paramPair: (Reference, ApiCall)) = {
-    val ref = paramPair._2.asReference
-    val t = null
-    val name = typeName(t, ref)
-    Map(COMMON -> Map(TYPE_NAME -> name, FIELDS -> fieldsForType(t), MEMBER_NAME -> memberName(t, ref)))
-  }
-}
-
-trait CommonDataStep extends EnrichmentStep[Type] with CommonData {
-
-  override def steps = types +: super.steps
-
-  private def avoidClashes(table: DenotationTable, name: String)
-                          (names: Iterable[String] = table.values.map(_ (COMMON)(TYPE_NAME).toString)): String =
-    if (names.exists(_ == name)) avoidClashes(table, name + "NameClash")(names) else name
-
-  /**
-    * Puts type information into the denotation table
-    * @return
-    */
-  protected def types: SingleStep = typeDef => table => typeDef match {
-    case (ref, t) =>
-      val name = avoidClashes(table, typeName(t, ref))()
-      Map(COMMON -> Map(TYPE_NAME -> name, FIELDS -> fieldsForType(t), MEMBER_NAME -> memberName(t, ref)))
-  }
-
-}
-
-trait CommonData {
-
-  def app: StrictModel
-
-  protected def typeName(t: Type, r: Reference, suffix: String = "") = t match {
-    case TypeRef(ref) =>
-      app.findType(ref) match {
-        case p: PrimitiveType => useType(p.name, suffix, "")
-        case _ => useType(ref, suffix, "")
-      }
-    case p: PrimitiveType => useType(t.name, suffix, "")
-    case _ => useType(r, suffix, "")
-  }
-
-  @tailrec
-  protected final def memberName(t: Type, r: Reference, suffix: String = ""): String = t match {
-    case TypeRef(ref) if ! app.findType(ref).isInstanceOf[TypeRef] => memberName(app.findType(ref), ref, suffix)
-    case TypeRef(ref) => useType(ref, suffix, "")
-    case p: PrimitiveType => useType(r, suffix, "")
-    case _ => useType(r, suffix, "")
-  }
-
-  protected def dereferenceFields(t: Composite): Seq[Field] =
-    t.descendants flatMap {
-      case td: TypeDef => td.fields
-      case r: TypeRef => app.findType(r.name) match {
-        case td: TypeDef => td.fields
-        case c: Composite => dereferenceFields(c)
-        case other =>
-          throw new IllegalStateException(s"Unexpected contents of Composite $r: $other")
-      }
-    }
-
-  protected def fieldsForType(t: Type): Seq[Field] = {
-    t match {
-      case c: Composite => dereferenceFields(c).distinct
-      case c: TypeDef => c.fields
-      case _ => Seq.empty[Field]
-    }
-  }
-
-  private def useType(ref: Reference, suffix: String, prefix: String) = {
-    val fullName = ref.qualifiedName(prefix, suffix)
-    fullName._2
-  }
-
-}
-trait ClassesStep extends EnrichmentStep[Type] {
-
-  override def steps = classes +: super.steps
-
-  /**
-    * Puts class related information into the denotation table
-    * @return
-    */
-  protected def classes: SingleStep = typeDef => table => typeDef match {
-    case (ref, t: TypeDef) if !ref.simple.contains("AllOf") && !ref.simple.contains("OneOf") =>
-      val traitName = app.discriminators.get(ref).map(_ => Map("name" -> typeNameDenotation(table, ref)))
-      Map("classes" -> (typeDefProps(ref, t)(table) + ("trait" -> traitName)))
-    case (ref, t: Composite) =>
-      Map("classes" -> (typeDefProps(ref, t)(table) + ("trait" -> t.root.map(r => Map("name" -> r.className)))))
-    case _ => empty
-  }
-
-  protected def typeDefProps(k: Reference, t: Type)(table: DenotationTable): Map[String, Any] = {
-    Map(
-      "name" -> typeNameDenotation(table, k),
-      "fields" -> typeFields(table, k).map { f =>
-        Map(
-          "name" -> escape(f.name.simple),
-          "type_name" -> typeNameDenotation(table, f.tpe.name)
-        )
-      }
-    )
-  }
-
-
-}
-
-trait TraitsStep extends EnrichmentStep[Type] {
-
-  override def steps = traits +: super.steps
-
-  /**
-    * Puts trait related information into the denotation table
-    * @return
-    */
-  protected def traits: SingleStep = typeDef => table => typeDef match {
-    case (ref, t: TypeDef) if app.discriminators.contains(ref) =>
-      Map("traits" -> typeDefProps(ref, t)(table))
-    case _ => empty
-  }
-
-  protected def typeDefProps(k: Reference, t: Type)(table: DenotationTable): Map[String, Any] // FIXME should be defined only once
-}
-
-trait AliasesStep extends EnrichmentStep[Type] {
-
-  override def steps = aliases +: super.steps
-
-  /**
-    * Puts type related information into the denotation table
-    * @return
-    */
-  protected val aliases: SingleStep = typeDef => table => typeDef match {
-    case (ref, t: Container) =>
-      Map("aliases" -> aliasProps(ref, t)(table))
-    case (k, v: PrimitiveType) =>
-      Map("aliases" -> mapForAlias(k, v)(table))
-    case (k, v: TypeRef) =>
-      Map("aliases" -> mapForAlias(v.name, v)(table))
-    case _ => empty
-  }
-
-  private def aliasProps(k: Reference, v: Container)(table: DenotationTable): Map[String, Any] = {
-    Map(
-      "name" -> typeNameDenotation(table, k),
-      "alias" -> v.alias,
-      "underlying_type" -> v.imports.headOption.map { _ => v.nestedTypes.map { t =>
-        typeNameDenotation(table, t.name)
-      }.mkString(", ")
-      }
-    )
-  }
-
-  private def mapForAlias(k: Reference, v: Type)(table: DenotationTable): Map[String, Object] = {
-    Map(
-      "name" -> memberNameDenotation(table, k),
-      "alias" -> typeNameDenotation(table, v.name),
-      "underlying_type" -> None
-    )
-  }
 }
 
 object DenotationNames {
